+++ src/js/getChord.test.js
@@ -1,0 +1,6 @@
+const getChord = require('./getChord').getChord
+
+test('return interval between notes', () => {
+    expect(getChord(['C','E','G'], [5,4]))
+        .toEqual('C Major')
+})
