+++ src/js/sound.js
@@ -1,10 +1,25 @@
 // const Howl = require('Howl')
-//let Howl = require('./external/Howler.min').Howl
+// let Howl = require('./external/Howler.min').Howl
 
 let preloaded = true
 
-
-
+function preload() {
+    let notes = []
+    //new Howl();
+    //preloading notes files
+    for (i = 0; i < 12; i++) {
+        notes[i] = new Howl({
+            src: [
+                'src/sound/mp3/' + i + '.mp3',
+                'src/sound/wav/' + i + '.wav'
+            ],
+            loop: false,
+            preload: true
+        });
+        //console.log('notes:' + notes[0])
+    }
+    return notes
+}
 
 
 //plays note when pressed/clicked
@@ -20,7 +35,8 @@ function playNote(NoteId, notes) {
     else {
         console.log('not preloaded')
     }
-    return notes[NoteId]
+    return notes
 }
 
 exports.playNote = playNote
+exports.preload = preload
