--- src/js/sound.t.js
@@ -1,6 +1,0 @@
-let playNote = require('./sound').playNote
-
-test('return chord name from chord notes', () => {
-    expect(playNote())
-        .toEqual('')
-})
